backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
google-generativeai>=0.3.2
google-genai>=0.1.0
pypdf>=4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]
bcrypt<4.0.0
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
faiss-cpu
numpy>=1.26.4
sentence-transformers>=2.3.1
rank_bm25>=0.2.2
nltk>=3.8.1
redis>=5.0.3
scikit-learn>=1.4.1.post1
slowapi>=0.1.9
